% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_names}
\alias{make_names}
\title{Make Syntactically Valid Names}
\usage{
make_names(names, unique = F, allow_ = T)
}
\arguments{
\item{names}{character vector to be coerced to syntactically valid names.}
}
\value{
A character vector of same length as names with each changed to a syntactically valid name,
in the current locale's encoding.
}
\description{
Just like \code{make.names} but with _ instead of .
}
