% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Mode}
\alias{Mode}
\title{Function for the mode of a sample}
\usage{
Mode(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
returns the mode(s) of the vector
}
\description{
Function for the mode of a sample
}
