% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{na_magic}
\alias{na_magic}
\title{Make NA magically turn into another value}
\usage{
na_magic(x, replace = 0)
}
\arguments{
\item{x}{vector with NAs}

\item{replace}{value for transmogrifying NAs into}
}
\value{
returns a vector of same length as x with transmogrified values
}
\description{
Make NA magically turn into another value
}
