% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prop_compare}
\alias{prop_compare}
\title{Prints proportion tables for a dataset and
training and testing subsets}
\usage{
prop_compare(df, y, train = training, test = testing)
}
\arguments{
\item{df}{full dataframe}

\item{y}{response variable to make proportion table for}

\item{train}{training dataset}

\item{test}{testing dataset}
}
\value{
prints three proportion tables
}
\description{
Prints proportion tables for a dataset and
training and testing subsets
}
