% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rescale}
\alias{rescale}
\title{Function to rescale a vector of numbers}
\usage{
rescale(x, a = 0, b = 1)
}
\arguments{
\item{x}{vector of values to be normalized}

\item{a}{minimum value for rescaling}

\item{b}{maximum value for rescaling}
}
\value{
returns a vector of values rescaled between a and b
}
\description{
Function to rescale a vector of numbers
}
