% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trim.leading}
\alias{trim.leading}
\title{Function to trim leading whitespace from a string}
\usage{
trim.leading(x)
}
\arguments{
\item{x}{string to trim whitespace from}
}
\value{
returns a string with leading whitespace trimmed
}
\description{
Function to trim leading whitespace from a string
}
