% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trim}
\alias{trim}
\title{Function to trim both leading and trailing whitespace from a string}
\usage{
trim(x)
}
\arguments{
\item{x}{string to trim whitespace from}
}
\value{
returns a string with both leading and trailing whitespace trimmed
}
\description{
Function to trim both leading and trailing whitespace from a string
}
