% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trim.trailing}
\alias{trim.trailing}
\title{Function to trim trailing whitespace from a string}
\usage{
trim.trailing(x)
}
\arguments{
\item{x}{string to trim whitespace from}
}
\value{
returns a string with trailing whitespace trimmed
}
\description{
Function to trim trailing whitespace from a string
}
